+++ open_ballot/js/controllers/ballots.js
@@ -1,0 +1,6 @@
+var controllers = angular.module('openBallotControllers');
+
+controllers.controller('ballotsController', ['$scope', 'api', function($scope, api) {
+  $scope.ballots = api.ballot_history.query();
+  $scope.text = 'hello ballots';
+}]);
