+++ open_ballot/js/controllers/contracts.js
@@ -1,0 +1,6 @@
+var controllers = angular.module('openBallotControllers');
+
+controllers.controller('contractsController', ['$scope', 'api', function($scope, api) {
+  $scope.contracts = api.contracts.query();
+  $scope.text = 'hello world';
+}]);
