+++ open_ballot/js/controllers/index.js
@@ -1,0 +1,5 @@
+var controllers = angular.module('openBallotControllers', ['openBallotServices']);
+
+require('./contracts');
+require('./ballot');
+require('./ballots');
