+++ open_ballot/js/directives/index.js
@@ -1,0 +1,1 @@
+var directives = angular.module('openBallotDirectives', []);
