+++ open_ballot/js/services/index.js
@@ -1,0 +1,3 @@
+var services = angular.module('openBallotServices', ['ngResource']);
+
+require('./api');
